move lib/cdo/properties.rb into pegasus/helpers/
