+++ bin/cron/analyze_hoc_activity
@@ -16,7 +16,7 @@ abort 'Script already running' unless only_one_running?(__FILE__)
 #
 require File.expand_path('../../../pegasus/src/env', __FILE__)
 require src_dir 'database'
-require 'cdo/properties'
+require_relative '../../pegasus/helpers/properties'
 require 'dynamic_config/dcdo'
 
 # Provides helper methods (analyze_day_fast and add_hashes) as well as the
