+++ bin/cron/geocode_hoc_activity
@@ -7,7 +7,7 @@ abort 'Script already running' unless only_one_running?(__FILE__)
 
 require File.expand_path('../../../pegasus/src/env', __FILE__)
 require src_dir 'database'
-require 'cdo/properties'
+require_relative '../../pegasus/helpers/properties'
 
 BATCH_SIZE = 5000
