+++ bin/cron/hoc_signup_counts
@@ -7,7 +7,7 @@ abort 'Script already running' unless only_one_running?(__FILE__)
 
 require File.expand_path('../../../pegasus/src/env', __FILE__)
 require 'cdo/db'
-require 'cdo/properties'
+require_relative '../../pegasus/helpers/properties'
 require 'dynamic_config/dcdo'
 
 DEFAULT_YEAR = 2017
