+++ bin/cron/update_census_mapbox
@@ -12,7 +12,7 @@ abort 'Script already running' unless only_one_running?(__FILE__)
 #   CDO.mapbox_upload_token: (the auth token for uploading tiles to Mapbox's servers)
 
 require_relative '../../dashboard/config/environment'
-require 'cdo/properties'
+require_relative '../../pegasus/helpers/properties'
 
 require_relative 'upload_to_mapbox'
