+++ bin/cron/update_hoc_map
@@ -14,7 +14,7 @@ abort 'Script already running' unless only_one_running?(__FILE__)
 
 require File.expand_path('../../../pegasus/src/env', __FILE__)
 require src_dir 'database'
-require 'cdo/properties'
+require_relative '../../pegasus/helpers/properties'
 require 'dynamic_config/dcdo'
 require 'open3'
 require 'tempfile'
