+++ bin/cron/update_project_count
@@ -10,7 +10,7 @@ abort 'Script already running' unless only_one_running?(__FILE__)
 
 require 'sequel'
 require File.expand_path('../../../pegasus/src/env', __FILE__)
-require 'cdo/properties'
+require_relative '../../pegasus/helpers/properties'
 require src_dir 'database'
 require_relative '../../deployment.rb'
 require 'cdo/honeybadger'
