+++ lib/analyze_hoc_activity_helper.rb
@@ -4,7 +4,7 @@
 #
 require File.expand_path('../../pegasus/src/env', __FILE__)
 require src_dir 'database'
-require 'cdo/properties'
+require_relative '../pegasus/helpers/properties'
 
 DASHBOARD_DB_READER = sequel_connect(CDO.dashboard_db_reader, CDO.dashboard_db_reader, query_timeout: 1200)
 PEGASUS_DB_READER = sequel_connect(CDO.pegasus_db_reader, CDO.pegasus_db_reader, query_timeout: 1200)
