+++ pegasus/helpers/properties.rb
