+++ pegasus/router.rb
@@ -8,7 +8,7 @@ require 'cdo/geocoder'
 require 'cdo/pegasus/graphics'
 require 'cdo/rack/cdo_deflater'
 require 'cdo/rack/request'
-require 'cdo/properties'
+require_relative 'helpers/properties'
 require 'cdo/languages'
 require 'dynamic_config/page_mode'
 require 'active_support'
